+++ src/main/java/br/com/livelo/orderflight/domain/dtos/pricing/response/PricingCalculateTaxes.java
@@ -9,6 +9,6 @@ import lombok.Data;
 @Builder
 public class PricingCalculateTaxes {
     private BigDecimal amount;
-    private Integer pointsAmount;
+    private BigDecimal pointsAmount;
     private String description;
 }
